Fix zip maker when cross-compiling
The zip maker needs to know the target platform when cross-compiling, rather than just using the host platform. This involved changing the maker interface, so it now takes an options hash to make interface changes like this easier in the future with respect to compatibility.

NOTE: This changes the maker interface, and will require any makers outside this repository to update their signatures.